начал делать сервис продуктов

+++ Entities/DataBase/Models/Product.cs
@@ -4,6 +4,6 @@ public class Product
 {
     public required string Name { get; set; }
     public required int Count { get; set; }
-    public required int Price { get; set; }
+    public required double Price { get; set; }
     public required ICollection<ProductCategory> Categories { get; set; }
 }
